Widen volumeHandle to long before shifting by 48 (CodeQL java/lshift-larger-than-type-width) (#231)
In IOMeter.dump() the composite event key was built as
"(volumeHandle << 48) + pageAddress". volumeHandle is an int, and Java uses
only the low 5 bits of an int shift distance, so "<< 48" was silently
evaluated as "<< 16" in 32-bit arithmetic before being widened to long. The
handle therefore placed volumeHandle in the wrong bit range (and could
overflow int), so events for different volumes/pages could collide on the
same HashMap key when dumping with analyzePages.

Cast volumeHandle to long first so the shift is performed in 64 bits:
((long) volumeHandle << 48) + pageAddress. This only affects the -a page
analysis of the journal dump utility.

+++ persistit/core/src/main/java/com/persistit/IOMeter.java
@@ -281,7 +281,7 @@ class IOMeter implements IOMeterMXBean {
                 if (analyzePages
                         && (op == WRITE_PAGE_TO_JOURNAL || op == READ_PAGE_FROM_JOURNAL || op == READ_PAGE_FROM_VOLUME
                                 || op == COPY_PAGE_FROM_JOURNAL || op == COPY_PAGE_TO_VOLUME || op == EVICT_PAGE_FROM_POOL)) {
-                    final long handle = (volumeHandle << 48) + pageAddress;
+                    final long handle = ((long) volumeHandle << 48) + pageAddress;
                     List<Event> list = events.get(handle);
                     if (list == null) {
                         list = new ArrayList<Event>(2);
